service/requirements.txt
fastapi
uvicorn
PyYAML
inflect
python-dotenv
sqlalchemy